Coremark stub for a barebone version of arty-a7.
This is merely a copy and setup of a barebone coremark (ie. no stdlib).
Compile with: `make CC=${RISCV_GCC} LD=${RISCV_GCC} link`. Notice the
use of GCC instead of LD, which is due to some problem with my version
of RISCV LD (could be some problem with multilib compiled `ld`).

IMPORTANT: Still missing code for text IO and timer!