UART model for Verilator testbench.
The model is to be used for simlations where text output is needed
(e.g. Coremark). The model shall be defined through SystemVerilog
defines (baud rate, data and stop bits) and received data are captured
in the C++ simulation engine, where it gets printed out at the end
of simulation.