Fixed dependencies in and target updates in Verilator testbench Makefile.
One problem was that re-running Verilator does not update timestamp of
the generated Makefile. Hence added a `touch` of that Makefile.

The other problem was a wrong dependency on the C++ testbench wrapper.
A change in that wrapper shall cause only re-running the build in the
generated Makefile and no need to re-run verilation.