OpenOCD support for Verilator testbench.
This commit adds support for connecting OpenOCD to the Verilator simulated
target. Details are described in `verilator/README.md`. The implementation
adds a remote bitbang server that runs as a thread of the Verilator simulation
and accepts connections from OpenOCD which it then translates to wiggles of
JTAG pins on the simulated target.

The Verilator C++ testbench was enhanced to allow command line options to
control simulation timeout and generation of VCD dumps. That is, those
functions are no longer controlled during compile time.